combine vBindShorthand.ts and types.ts
