--- packages/language-core/lib/codegen/types.ts
@@ -1,9 +1,0 @@
-export interface InlayHintInfo {
-	blockName: string;
-	offset: number;
-	setting: string;
-	label: string;
-	tooltip?: string;
-	paddingRight?: boolean;
-	paddingLeft?: boolean;
-}
